fix: 동시성 테스트 수정
