+++ src/test/kotlin/com/reservation/ticket/concert/interfaces/api/user/UserControllerTest.kt
@@ -1,54 +1,167 @@
 package com.reservation.ticket.concert.interfaces.api.user
 
 import com.reservation.ticket.concert.application.service.QueueService
+import com.reservation.ticket.concert.application.service.UserService
+import com.reservation.ticket.concert.domain.Concert
+import com.reservation.ticket.concert.domain.User
+import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
-
-import org.mockito.Mockito.`when`
+import org.junit.jupiter.api.extension.ExtendWith
 import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest
-import org.springframework.boot.test.mock.mockito.MockBean
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc
+import org.springframework.boot.test.context.SpringBootTest
+import org.springframework.http.MediaType
+import org.springframework.test.context.junit.jupiter.SpringExtension
 import org.springframework.test.web.servlet.MockMvc
-import org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get
-import org.springframework.test.web.servlet.result.MockMvcResultMatchers.content
-import org.springframework.test.web.servlet.result.MockMvcResultMatchers.status
-import java.util.*
+import org.springframework.test.web.servlet.get
+import org.springframework.test.web.servlet.put
+import java.time.LocalDateTime
+import java.util.UUID
+import java.util.concurrent.CountDownLatch
+import java.util.concurrent.ExecutorService
+import java.util.concurrent.Executors
 
-@WebMvcTest(UserController::class)
+@ExtendWith(SpringExtension::class)
+@SpringBootTest
+@AutoConfigureMockMvc
 class UserControllerTest {
 
     @Autowired
     private lateinit var mockMvc: MockMvc
 
-    @MockBean
+    @Autowired
     private lateinit var queueService: QueueService
 
+    @Autowired
+    private lateinit var userService: UserService
+
     @Test
-    fun `getUserQueuePosition should return position successfully`() {
-        // Given
-        val userId = UUID.randomUUID()
-        val position = 3
+    fun `multiple users send requests to get their queue position`() {
 
-        // Mocking the service
-        `when`(queueService.getUserQueuePosition(userId)).thenReturn(position)
+        val numberOfThreads = 10
+        val executorService: ExecutorService = Executors.newFixedThreadPool(numberOfThreads)
+        val latch = CountDownLatch(numberOfThreads)
 
-        // When & Then
-        mockMvc.perform(get("/queue/$userId/position"))
-            .andExpect(status().isOk)
-            .andExpect(content().string("3"))
+        val tempConcert = Concert(name = "Concert 1", location = "lollapalooza", date = LocalDateTime.now(), availableTickets = 48)
+        val users = (1..10).map {
+            userService.saveUser(User(UUID.randomUUID(),"testUser$it", "password$it"))
+        }
+
+        val userIds = users.map { it -> it.id }
+        val expectedPositions = (1..numberOfThreads).toList()
+
+        val responses = mutableListOf<String>()
+
+        for (i in 0 until numberOfThreads) {
+            val userId = userIds[i]
+            executorService.submit {
+                try {
+                    val response = mockMvc.get("/api/$userId/position")
+                        .andExpect { status { isOk() } }
+                        .andReturn().response.contentAsString
+
+                    synchronized(responses) {
+                        responses.add(response)
+                    }
+                } finally {
+                    latch.countDown()
+                }
+            }
+        }
+
+        latch.await()
+
+        assertThat(responses.size).isEqualTo(numberOfThreads)
+
+        expectedPositions.forEach { position ->
+            assertThat(responses).contains(position.toString())
+        }
     }
 
     @Test
-    fun `getUserQueuePosition should return 400 when user not in queue`() {
-        // Given
+    fun `multiple users send their own point GET requests`() {
+
+        val numberOfThreads = 10
+        val executorService: ExecutorService = Executors.newFixedThreadPool(numberOfThreads)
+        val latch = CountDownLatch(numberOfThreads)
+
         val userId = UUID.randomUUID()
 
-        // Mocking the service to throw an exception
-        `when`(queueService.getUserQueuePosition(userId))
-            .thenThrow(IllegalArgumentException("해당 유저는 대기열에 없습니다."))
+        val responses = mutableListOf<String>()
 
-        // When & Then
-        mockMvc.perform(get("/queue/$userId/position"))
-            .andExpect(status().isBadRequest)
-            .andExpect(content().string("해당 유저는 대기열에 없습니다."))
+        for (i in 0 until numberOfThreads) {
+            executorService.submit {
+                try {
+                    // 포인트 조회 요청
+                    val response = mockMvc.get("/api/user/$userId/point") {
+                        accept(MediaType.APPLICATION_JSON)
+                    }.andExpect {
+                        status { isOk() }
+                    }.andReturn().response.contentAsString
+
+                    synchronized(responses) {
+                        responses.add(response)
+                    }
+                } finally {
+                    latch.countDown()
+                }
+            }
+        }
+
+        latch.await()
+
+        assertThat(responses.size).isEqualTo(numberOfThreads)
+
+        val firstResponse = responses.first()
+        responses.forEach { response ->
+            assertThat(response).isEqualTo(firstResponse)
+        }
     }
+    @Test
+    fun `multiple users send 'add points' requests at once`() {
+
+        val numberOfThreads = 10
+        val executorService: ExecutorService = Executors.newFixedThreadPool(numberOfThreads)
+        val latch = CountDownLatch(numberOfThreads)
+
+        val userId = 1L
+
+        val request = """
+            {
+                "amount": 100
+            }
+        """.trimIndent()
+
+        val responses = mutableListOf<String>()
+
+        for (i in 0 until numberOfThreads) {
+            executorService.submit {
+                try {
+                    // 포인트 추가 요청
+                    val response = mockMvc.put("/api/user/$userId/point/add") {
+                        contentType = MediaType.APPLICATION_JSON
+                        content = request
+                    }.andExpect {
+                        status { isOk() }
+                    }.andReturn().response.contentAsString
+
+                    synchronized(responses) {
+                        responses.add(response)
+                    }
+                } finally {
+                    latch.countDown()
+                }
+            }
+        }
+
+        latch.await()
+
+        assertThat(responses.size).isEqualTo(numberOfThreads)
+
+        responses.forEach { response ->
+            assertThat(response).contains("\"code\":\"OK\"")
+            assertThat(response).contains("\"point\":100")
+        }
+    }
+
 }
